Refactor configuration types and update .gitignore
