+++ src/types/config.d.ts
@@ -7,3 +7,22 @@ declare module "config" {
   const config: IConfig;
   export default config;
 }
+
+export interface AppConfig {
+  name: string;
+  version: string;
+  description: string;
+}
+
+export interface CommandConfig {
+  name: string;
+  description: string;
+  module?: string;
+  subcommands?: boolean;
+  children?: CommandConfig[];
+}
+
+export interface Config {
+  app: AppConfig;
+  commands: CommandConfig[];
+}
